requirements.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
black==19.10b0
click==7.1.2
colorama==0.4.3
Django==3.0.6
django-crispy-forms==1.9.1
flake8==3.8.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.8.0
Pillow==7.1.2
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
regex==2020.5.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
